refactor: 즐겨찾기 로직 분리
